requirements.txt
# Audio/Video processing
ffmpeg-python>=0.2.0
pydub>=0.25.1
librosa>=0.10.0

# Whisper (ASR)
openai-whisper>=20231117

# Translation
transformers>=4.30.0
sentencepiece>=0.1.99

# TTS Vietnamese (Edge TTS - no compilation needed)
edge-tts>=6.1.0

# Deep Learning (CPU)
torch>=2.0.0
torchaudio>=2.0.0

# Utilities
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Progress bar & CLI
tqdm>=4.65.0
